Move nagivation logic to the screen
